decode ethereum transaction independent enviorment struct
