+++ transfer/root.go
@@ -3,7 +3,6 @@ package transfer
 import (
 	"encoding/hex"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	"math"
@@ -11,9 +10,7 @@ import (
 	"net/http"
 	"os"
 	"reflect"
-	"regexp"
 	"runtime/debug"
-	"strconv"
 	"strings"
 	"time"
 
@@ -29,14 +26,11 @@ import (
 	"github.com/Conflux-Chain/go-conflux-sdk/types/cfxaddress"
 	sdkerrors "github.com/Conflux-Chain/go-conflux-sdk/types/errors"
 	"github.com/Conflux-Chain/go-conflux-sdk/utils"
-	"github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/rlp"
 	clientRpc "github.com/openweb3/go-rpc-provider"
 
 	"github.com/shopspring/decimal"
 
-	"github.com/Conflux-Chain/go-conflux-sdk/middleware"
 	"github.com/spf13/cobra"
 )
 
@@ -57,7 +51,7 @@ var (
 	weight           decimal.Decimal
 
 	perBatchNum uint
-	env         enviorment
+	env         Enviorment
 )
 
 func init() {
@@ -75,13 +69,11 @@ func init() {
 
 	rootCmd.PersistentFlags().UintVar(&perBatchNum, "batch", 1000, "send tx number per batch")
 	formatReceiverNumber(sentWeightInStr)
-
-	getProcessEnv()
 }
 
 func doTransfers(cmd *cobra.Command, args []string) {
 	fmt.Println("Initial enviorment")
-	initialEnviorment()
+	env = *NewEnviorment()
 
 	defer clearCacheFile()
 
@@ -90,7 +82,7 @@ func doTransfers(cmd *cobra.Command, args []string) {
 
 	// list cfx and ctoken for user select
 	tokenSymbol, tokenAddress := selectToken()
-	fmt.Printf("Selected token: %v, contract address: %v\n", tokenSymbol, addrDisplay(tokenAddress))
+	fmt.Printf("Selected token: %v, contract address: %v\n", tokenSymbol, env.AddrDisplay(tokenAddress))
 
 	// transfer
 	fmt.Println("===== Start batch transfer =====")
@@ -176,7 +168,7 @@ func estimateGasAndCollateral(tokenAddress *cfxaddress.Address) types.Estimate {
 		_doubledGasLimit = big.NewInt(15000000)
 	}
 	em.GasLimit = types.NewBigIntByRaw(_doubledGasLimit)
-	logrus.Debugf("estimate <from %v, to %v, data %v> result: gas %v, collateral %v", addrDisplay(callReq.From), addrDisplay(callReq.To), *callReq.Data, em.GasLimit, em.StorageCollateralized)
+	logrus.Debugf("estimate <from %v, to %v, data %v> result: gas %v, collateral %v", env.AddrDisplay(callReq.From), env.AddrDisplay(callReq.To), *callReq.Data, em.GasLimit, em.StorageCollateralized)
 	return em
 }
 
@@ -191,7 +183,7 @@ func creatOneBatchElems(oneBatchReceiver []Receiver, tokenAddress *cfxaddress.Ad
 
 		receiver := cfxaddress.MustNew(v.Address, env.networkID)
 		fmt.Printf("%v. Sign send %v to %v with value %v nonce %v done\n",
-			startCnt+i, tokenSymbol, addrDisplay(&receiver),
+			startCnt+i, tokenSymbol, env.AddrDisplay(&receiver),
 			util.DisplayValueWithUnit(calcValue(weight, v.AmountInCfx)), tx.Nonce)
 		env.nonce = env.nonce.Add(env.nonce, big.NewInt(1))
 	}
@@ -207,7 +199,7 @@ func createBatchElemItem(tx *types.UnsignedTransaction) clientRpc.BatchElem {
 
 	// sign
 	// encoded, err := env.am.SignTransaction(*tx)
-	encoded, err := signTx(tx)
+	encoded, err := env.SignTx(tx)
 	util.OsExitIfErr(err, "Failed to sign transaction %+v", tx)
 
 	// fmt.Printf("%v. Sign send %v to %v with value %v nonce %v done\n",
@@ -221,18 +213,6 @@ func createBatchElemItem(tx *types.UnsignedTransaction) clientRpc.BatchElem {
 		Args:   []interface{}{"0x" + hex.EncodeToString(encoded)},
 		Result: batchElemResult,
 	}
-}
-
-func signTx(tx *types.UnsignedTransaction) ([]byte, error) {
-	switch env.space {
-	case types.SPACE_NATIVE:
-		return env.am.SignTransaction(*tx)
-	case types.SPACE_EVM:
-		eTx, addr, chainID := adaptToEthTx(tx)
-		eTx = signEthLegacyTx(env.ethKeystore, addr, eTx, chainID)
-		return rlp.EncodeToBytes(eTx)
-	}
-	return nil, errors.New("unkown space")
 }
 
 func batchSendAndWaitReceipt(rpcBatchElems []clientRpc.BatchElem) BatchSummary {
@@ -398,7 +378,7 @@ func improveTxGasPrice(be clientRpc.BatchElem) clientRpc.BatchElem {
 
 	signedTx := &types.SignedTransaction{}
 	err = signedTx.Decode(rawTxByts, env.networkID)
-	util.OsExitIfErr(err, "%v. Failed to decode signed tx %v\n")
+	util.OsExitIfErr(err, "Failed to decode signed tx %v", signedTx)
 
 	logrus.Debugf("before gasPrice %v\n", signedTx.UnsignedTransaction.GasPrice.ToInt())
 	signedTx.UnsignedTransaction.GasPrice = types.NewBigIntByRaw(new(big.Int).Add(signedTx.UnsignedTransaction.GasPrice.ToInt(), big.NewInt(1000)))
@@ -565,7 +545,7 @@ func getTransferData(contractAddress types.Address, reciever types.Address, amou
 	ctoken := common.MustGetCTokenContract(contractAddress.String())
 	// data, err := ctoken.GetData("send", reciever.MustGetCommonAddress(), amountInDrip, []byte{})
 	data, err := ctoken.GetData("transfer", reciever.MustGetCommonAddress(), amountInDrip)
-	util.OsExitIfErr(err, "Failed to get data of transfer ctoken %v to %v amount %v", addrDisplay(&contractAddress), addrDisplay(&reciever), amountInDrip)
+	util.OsExitIfErr(err, "Failed to get data of transfer ctoken %v to %v amount %v", env.AddrDisplay(&contractAddress), env.AddrDisplay(&reciever), amountInDrip)
 	return data
 }
 
@@ -579,7 +559,7 @@ func selectToken() (symbol string, contractAddress *types.Address) {
 		return "", nil
 	}
 
-	url := getTokenListUrl()
+	url := env.GetTokenListUrl()
 	req, _ := http.NewRequest("GET", url, nil)
 
 	res, err := http.DefaultClient.Do(req)
@@ -602,7 +582,7 @@ func selectToken() (symbol string, contractAddress *types.Address) {
 	fmt.Println("\nThese are the token list you could batch transfer:")
 	fmt.Printf("%v. Token: %v\n", 1, "CFX")
 	for i := range tokenList.List {
-		fmt.Printf("%v. Token: %v, Contract Address: %v\n", i+2, tokenList.List[i].Symbol, addrDisplay(&tokenList.List[i].Address))
+		fmt.Printf("%v. Token: %v, Contract Address: %v\n", i+2, tokenList.List[i].Symbol, env.AddrDisplay(&tokenList.List[i].Address))
 	}
 
 	selectedIdx := getSelectedIndex(len(tokenList.List) + 2)
@@ -665,59 +645,6 @@ func getSelectedIndex(tokensCount int) int {
 	return selectedIdx
 }
 
-func inputPassword() string {
-	if env.isDebugMode {
-		return "123"
-	}
-	return account.MustInputPassword("Enter password: ")
-}
-
-// ================= inits ================================
-func initialEnviorment() {
-	env.am = account.DefaultAccountManager
-	env.ethKeystore = keystore.NewKeyStore("keystore", keystore.StandardScryptN, keystore.StandardScryptP)
-
-	env.client = rpc.MustCreateClientWithRetry(100)
-	env.client.SetAccountManager(env.am)
-
-	// fmt.Printf("env.logLevel: %v\n", env.logLevel)
-	if env.logLevel >= logrus.DebugLevel {
-		env.client.UseCallRpcMiddleware(middleware.CallRpcConsoleMiddleware)
-		env.client.UseBatchCallRpcMiddleware(middleware.BatchCallRpcConsoleMiddleware)
-	}
-
-	status, err := env.client.GetStatus()
-	util.OsExitIfErr(err, "Failed to get status")
-
-	env.chainID = uint32(status.ChainID)
-	env.networkID = uint32(status.NetworkID)
-	env.space = getSpace()
-
-	env.pState = loadProcessState()
-	env.from = cfxaddress.MustNew(account.MustParseAccount().GetHexAddress(), env.networkID)
-	env.fromEspace = mustGetAccount(env.ethKeystore, env.from.MustGetCommonAddress()).Address
-	// env.pState.refreshSpaceAndSave(types.SpaceType(space))
-	env.pState.refreshSenderAndSave(&env.from)
-
-	password := inputPassword()
-	err = env.am.Unlock(env.from, password)
-	util.OsExitIfErr(err, "Failed to unlock account")
-	ethAcc := mustGetAccount(env.ethKeystore, env.from.MustGetCommonAddress())
-	err = env.ethKeystore.Unlock(ethAcc, password)
-	util.OsExitIfErr(err, "Failed to unlock account")
-
-	fmt.Printf("Account %v is unlocked\n", addrDisplay(&env.from))
-
-	// get inital Nonce
-	_nonce, e := env.client.TxPool().NextNonce(*getFromAddress())
-	env.nonce = _nonce.ToInt()
-	util.OsExitIfErr(e, "Failed to get nonce of from %v", addrDisplay(&env.from))
-
-	epoch, err := env.client.GetEpochNumber(types.EpochLatestState)
-	util.OsExitIfErr(err, "Failed to get epoch")
-	env.epochHeight = epoch.ToInt().Uint64()
-}
-
 func formatReceiverNumber(sentWeightInStr string) {
 	var err error
 	weight, err = decimal.NewFromString(sentWeightInStr)
@@ -727,65 +654,6 @@ func formatReceiverNumber(sentWeightInStr string) {
 // SetParent sets parent command
 func SetParent(parent *cobra.Command) {
 	parent.AddCommand(rootCmd)
-}
-
-func mustParseReceivers() []Receiver {
-	// read csv file
-	content, err := ioutil.ReadFile(receiverListFile)
-	util.OsExitIfErr(err, "Failed to read file %v", receiverListFile)
-
-	// parse to struct
-	lines := strings.Split(string(content), "\n")
-	receiverInfos := []Receiver{}
-
-	invalids := []string{}
-
-	for i, v := range lines {
-		v = strings.Replace(v, "\t", " ", -1)
-		v = strings.Replace(v, ",", " ", -1)
-		items := strings.Fields(v)
-		if len(v) == 0 {
-			continue
-		}
-
-		if len(items) != 2 {
-			util.OsExit("Line %v: %#v column number is %v, which shoule be 2\n", i, v, len(items))
-		}
-
-		isDecimal, err := regexp.Match(`^\d+\.?\d*$`, []byte(items[1]))
-		util.OsExitIfErr(err, "Failed to regex check %v ", items[1])
-		if !isDecimal {
-			invalids = append(invalids, fmt.Sprintf("Line %v: Number %v is unsupported, only supports pure number format, scientific notation like 1e18 and other representation format are unspported", i+1, items[1]))
-			continue
-		}
-
-		amountInCfx, err := decimal.NewFromString(items[1])
-		if err != nil {
-			invalids = append(invalids, fmt.Sprintf("Line %v: Failed to parse %v to int, errmsg:%v", i+1, items[1], err.Error()))
-			continue
-		}
-
-		_, err = cfxaddress.New(items[0])
-		if err != nil {
-			invalids = append(invalids, fmt.Sprintf("Line %v: Failed to create cfx address by %v, Errmsg: %v", i+1, items[0], err.Error()))
-			continue
-		}
-
-		info := Receiver{
-			Address:     items[0],
-			AmountInCfx: amountInCfx,
-		}
-
-		receiverInfos = append(receiverInfos, info)
-	}
-
-	if len(invalids) > 0 {
-		errMsg := fmt.Sprintf("Invalid Recevier info exists:\n%v", strings.Join(invalids, "\n"))
-		util.OsExit(errMsg)
-	}
-
-	fmt.Printf("Receiver list count :%+v\n", len(receiverInfos))
-	return receiverInfos
 }
 
 func checkBalance(client *sdk.Client, from types.Address, receivers []Receiver, token *types.Address, tokenSymbol string) {
@@ -802,21 +670,21 @@ func checkBalance(client *sdk.Client, from types.Address, receivers []Receiver, 
 		storageNeed *big.Int = big.NewInt(0)
 	)
 
-	_balance, err := client.GetBalance(*getFromAddress())
+	_balance, err := client.GetBalance(*env.GetFromAddressOfSpace())
 	cfxBalance = _balance.ToInt()
-	util.OsExitIfErr(err, "Failed to get CFX balance of %v", addrDisplay(&from))
+	util.OsExitIfErr(err, "Failed to get CFX balance of %v", env.AddrDisplay(&from))
 
 	if token != nil {
 		contract := common.MustGetCTokenContract(token.String())
 		err := contract.Call(nil, &tokenBalance, "balanceOf", from.MustGetCommonAddress())
-		util.OsExitIfErr(err, "Failed to get token %v balance of %v", tokenSymbol, addrDisplay(&from))
+		util.OsExitIfErr(err, "Failed to get token %v balance of %v", tokenSymbol, env.AddrDisplay(&from))
 
 		_price := (*hexutil.Big)(account.MustParsePrice())
 		em := estimateGasAndCollateral(token)
 		aginstResp, err := client.CheckBalanceAgainstTransaction(from, *token, em.GasLimit, _price, em.StorageCollateralized)
 		util.OsExitIfErr(err, "Failed to check balance against tx")
 
-		logrus.Debugf("CheckBalanceAgainstTransaction from %v gaslimit %v gasprice %v storage collateral %v : %+v", addrDisplay(&from), em.GasLimit, _price, em.StorageCollateralized, aginstResp)
+		logrus.Debugf("CheckBalanceAgainstTransaction from %v gaslimit %v gasprice %v storage collateral %v : %+v", env.AddrDisplay(&from), em.GasLimit, _price, em.StorageCollateralized, aginstResp)
 
 		// needPayGas = aginstResp.WillPayTxFee
 		// needPayStorage = aginstResp.WillPayCollateral
@@ -847,7 +715,7 @@ func checkBalance(client *sdk.Client, from types.Address, receivers []Receiver, 
 		if cfxBalance.Cmp(cfxNeed) < 0 {
 			// clearCacheFile()
 			msg := fmt.Sprintf("Balance of %v is not enough,  need %v, has %v",
-				addrDisplay(&from), util.DisplayValueWithUnit(cfxNeed), util.DisplayValueWithUnit(cfxBalance))
+				env.AddrDisplay(&from), util.DisplayValueWithUnit(cfxNeed), util.DisplayValueWithUnit(cfxBalance))
 			util.OsExit(msg)
 		}
 	} else {
@@ -856,7 +724,7 @@ func checkBalance(client *sdk.Client, from types.Address, receivers []Receiver, 
 			// clearCacheFile()
 			msg := fmt.Sprintf("Token %v balance of %v is not enough or CFX balance not enough to pay gas,"+
 				"%v need %v, has %v, CFX need %v, has %v",
-				tokenSymbol, addrDisplay(&from),
+				tokenSymbol, env.AddrDisplay(&from),
 				tokenSymbol, util.DisplayValueWithUnit(receiveNeed, tokenSymbol), util.DisplayValueWithUnit(tokenBalance, tokenSymbol),
 				util.DisplayValueWithUnit(cfxNeed), util.DisplayValueWithUnit(cfxBalance),
 			)
@@ -868,92 +736,19 @@ func checkBalance(client *sdk.Client, from types.Address, receivers []Receiver, 
 		tokenBalance = cfxBalance
 	}
 
-	fmt.Printf("Balance of %v is enough, %v need %v, fee need %v; token has %v, cfx has %v\n", addrDisplay(&from), tokenSymbol,
+	fmt.Printf("Balance of %v is enough, %v need %v, fee need %v; token has %v, cfx has %v\n", env.AddrDisplay(&from), tokenSymbol,
 		util.DisplayValueWithUnit(receiveNeed, tokenSymbol), util.DisplayValueWithUnit(new(big.Int).Add(gasNeed, storageNeed)),
 		util.DisplayValueWithUnit(tokenBalance, tokenSymbol), util.DisplayValueWithUnit(cfxBalance))
 }
 
 func exitIfHasPendingTxs() {
 	// exit if has pending tx
-	nonce, err := env.client.GetNextNonce(*getFromAddress())
+	nonce, err := env.client.GetNextNonce(*env.GetFromAddressOfSpace())
 	util.OsExitIfErr(err, "Failed to get account next nonce")
-	pendingNonce, err := env.client.TxPool().NextNonce(*getFromAddress())
+	pendingNonce, err := env.client.TxPool().NextNonce(*env.GetFromAddressOfSpace())
 	util.OsExitIfErr(err, "Failed to get account pending nonce")
 	if nonce.ToInt().Cmp(pendingNonce.ToInt()) < 0 {
-		fmt.Printf("Exit, account %v has pending txs, please clear it first\n", addrDisplay(&env.from))
+		fmt.Printf("Exit, account %v has pending txs, please clear it first\n", env.AddrDisplay(&env.from))
 		os.Exit(0)
 	}
-}
-
-func getProcessEnv() {
-	modeStr := os.Getenv("MODE")
-	if modeStr == "DEBUG" || modeStr == "debug" {
-		env.isDebugMode = true
-	}
-
-	env.logLevel = logrus.InfoLevel
-	logLevelStr := os.Getenv("LOG_LEVEL")
-	if logLevelStr != "" {
-		logLevel, err := strconv.Atoi(logLevelStr)
-		if err != nil {
-			util.OsExitIfErr(err, "Failed to get log level")
-		}
-		env.logLevel = logrus.Level(logLevel)
-	}
-}
-
-func getTokenListUrl() string {
-	switch env.networkID {
-	case util.MAINNET:
-		return "https://confluxscan.io/v1/token?orderBy=transferCount&reverse=true&skip=0&limit=100&fields=price"
-	case util.TESTNET:
-		return "https://testnet.confluxscan.io/v1/token?orderBy=transferCount&reverse=true&skip=0&limit=100&fields=price"
-	case util.ESPACE_MAINNET:
-		return "https://evm.confluxscan.io/v1/token?orderBy=transferCount&reverse=true&skip=0&limit=100&fields=price"
-	case util.ESPACE_TESTNET:
-		return "https://evmtestnet.confluxscan.io/v1/token?orderBy=transferCount&reverse=true&skip=0&limit=100&fields=price"
-	}
-	panic("unknown network")
-}
-
-func getSpace() types.SpaceType {
-	switch env.networkID {
-	case util.MAINNET:
-		return types.SPACE_NATIVE
-	case util.TESTNET:
-		return types.SPACE_NATIVE
-	case util.ESPACE_MAINNET:
-		return types.SPACE_EVM
-	case util.ESPACE_TESTNET:
-		return types.SPACE_EVM
-	}
-	panic("unknown network")
-}
-
-func addrDisplay(addr *cfxaddress.Address) string {
-	if addr == nil {
-		return "nil"
-	}
-
-	switch env.space {
-	case types.SPACE_NATIVE:
-		return addr.String()
-	case types.SPACE_EVM:
-		if addr.String() == env.from.String() {
-			return env.fromEspace.String()
-		}
-		return addr.MustGetCommonAddress().String()
-	}
-	panic("unknown space")
-}
-
-func getFromAddress() *cfxaddress.Address {
-	switch env.space {
-	case types.SPACE_NATIVE:
-		return &env.from
-	case types.SPACE_EVM:
-		addr := cfxaddress.MustNewFromCommon(env.fromEspace, env.networkID)
-		return &addr
-	}
-	panic("unknown space")
 }
